% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataspec.R
\docType{data}
\name{dataspec_p4}
\alias{dataspec_p4}
\alias{dataspec_p5}
\alias{dataspec_pd}
\title{Specific absorption coefficients for PROSPECT model}
\format{An object of class \code{matrix} with 2101 rows and 3 columns.}
\usage{
dataspec_p4

dataspec_p5

dataspec_pd
}
\description{
Specific absorption coefficients for PROSPECT model
}
\keyword{datasets}
