% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidf.R
\name{dcum}
\alias{dcum}
\title{Leaf inclination distribution function from Verhoef dissertation}
\usage{
dcum(a, b, theta)
}
\arguments{
\item{a, b}{LIDF parameters}

\item{theta}{Angle}
}
\value{
LIDF fraction value
}
\description{
Leaf inclination distribution function from Verhoef dissertation
}
