% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpm.R
\name{e1_approx}
\alias{e1_approx}
\title{Swami and Ohija approximation to the Exponential Integral. This is very
simple, but differences with more precise (but expensive) approaches are
almost completely washed out by PROSPECT math.}
\usage{
e1_approx(x)
}
\description{
Swami and Ohija approximation to the Exponential Integral. This is very
simple, but differences with more precise (but expensive) approaches are
almost completely washed out by PROSPECT math.
}
