% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edr.R
\name{edr_r}
\alias{edr_r}
\title{Run ED two-stream radiative transfer model}
\usage{
edr_r(
  pft,
  lai,
  wai,
  cai,
  N,
  Cab,
  Car,
  Cw,
  Cm,
  orient_factor,
  clumping_factor,
  soil_moisture,
  direct_sky_frac,
  czen,
  wood_reflect = matrix(rep(wood_spec, max(pft)), 2101),
  wavelengths = seq(400, 2500)
)
}
\arguments{
\item{pft}{PFT identities of each cohort, as integer (ncoh)}

\item{lai}{Leaf area index of each cohort (ncoh)}

\item{wai}{Wood area index of each cohort (ncoh)}

\item{cai}{Crown area of each cohort (ncoh)}

\item{N}{PROSPECT 5 effective number of leaf mesophyll layers (>1; npft)}

\item{Cab}{PROSPECT 5 leaf chlorophyll content (ug cm-2; npft)}

\item{Car}{PROSPECT 5 leaf carotenoid content (ug cm-2; npft)}

\item{Cw}{PROSPECT 5 leaf water content (g cm-2; npft)}

\item{Cm}{PROSPECT 5 leaf dry matter content (g cm-2; npft)}

\item{orient_factor}{Orient factor (npft)}

\item{clumping_factor}{Clumping factor (npft)}

\item{soil_moisture}{Soil moisture index (0-1; 0 = completely dry,
1 = completely wet)}

\item{direct_sky_frac}{Fraction of incident solar radiation that is
direct (0-1; 0 = all diffuse radiation)}

\item{czen}{Cosine of angle of incidence (\code{cosaio})}

\item{wood_reflect}{Wood reflectance spectrum. Default \code{\link[=wood_spectrum]{wood_spectrum()}}}

\item{wavelengths}{Numeric vector of wavelengths to use, in nm
(nwl). Default is 400:2500.}
}
\description{
Wrapper around \link{sw_two_stream} that also calls PROSPECT 5 (see
\code{\link[PEcAnRTM:prospect]{PEcAnRTM::prospect()}}) to generate the leaf reflectance and
transmittance spectra, \code{\link[=hapke_soil]{hapke_soil()}} to generate the soil spectrum
given \code{soil_moisture}, and a "flat" incident solar spectra for
direct and diffuse light based on \code{direct_sky_frac}.
}
\author{
Alexey Shiklomanov
}
