% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eq}
\alias{eq}
\title{Test for equality within machine tolerance}
\usage{
eq(x, y)
}
\arguments{
\item{x, y}{Numeric values to compare}
}
\value{
\code{TRUE} if equal, \code{FALSE} otherwise
}
\description{
Test for equality within machine tolerance
}
\author{
Alexey Shiklomanov
}
