% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foursail.R
\name{foursail}
\alias{foursail}
\title{4SAIL model of canopy reflectance}
\usage{
foursail(
  leaf_refl,
  leaf_trans,
  soil_refl,
  LAI,
  hot_spot = 0,
  LIDFa = -0.35,
  LIDFb = -0.15,
  solar_zenith = 0,
  instrument_zenith = 0,
  azimuth = 0
)
}
\value{

}
\description{
4SAIL model of canopy reflectance
}
\author{
Alexey Shiklomanov
}
