% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpm.R
\name{gpm}
\alias{gpm}
\title{Generalized plate model}
\usage{
gpm(k, N, talf, t12, t21)
}
\arguments{
\item{k}{Optical depth (\code{numeric(nwl)})}

\item{N}{Effective number of leaf layers (\code{numeric(1)})}

\item{talf, t12, t21}{Pre-calculated quantities based on refractive index and angle}
}
\value{
A length-2 list containing the modeled reflectance and transmittance
}
\description{
Generalized plate model
}
\references{
Allen W.A., Gausman H.W., Richardson A.J., Thomas J.R.
(1969), Interaction of isotropic ligth with a compact plant leaf,
Journal of the Optical Society of American, 59:1376-1379.

Stokes G.G. (1862), On the intensity of the light
reflected from or transmitted through a pile of plates,
Proceedings of the Royal Society of London, 11:545-556.
}
\author{
Alexey Shiklomanov
}
