% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hapke-soil.R
\name{hapke_soil}
\alias{hapke_soil}
\title{Simple model of soil reflectance as a function of soil moisture}
\usage{
hapke_soil(soil_moisture)
}
\arguments{
\item{soil_moisture}{Soil moisture index (0-1; 0 = completely dry,
1 = completely wet)}
}
\value{
Spectrum of soil reflectance (400:2500 nm)
}
\description{
Just the average of \link{wet_soil} and \link{dry_soil}, weighted by
\code{soil_moisture}. \code{soil_moisture} of 0 means fully dry soil, and
}
\author{
Alexey Shiklomanov
}
