% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liberty.R
\name{liberty}
\alias{liberty}
\title{LIBERTY model}
\usage{
liberty(
  D,
  xu,
  thick,
  baseline,
  element,
  c_factor,
  w_factor,
  l_factor,
  p_factor
)
}
\arguments{
\item{D}{Cell diameter. Average leaf cell diameter (m-6). Typical value 40 (20-100).}

\item{xu}{Intercellular air space. Determinant for the amount of radiative
flux passing between cells. Typical value 0.045 (0.01-0.1)}

\item{thick}{Leaf thickness. Arbitrary value to determine single leaf
reflectance and transmittance from infinite reflectance criteria. Typical
value 1.6 (1-10).}

\item{baseline}{Baseline absorption. Wavelength-independent absorption to
compensate for changes in absolute reflectance. Typical value 0.0006 for
fresh leaves or or 0.0004 for dry leaves.}

\item{element}{Albino absorption. Absorption in the visible region due to
lignin. Typical value 2 (0-4).}

\item{c_factor}{Chlorophyll content (mg m-2). Typical value 200 (0-600).}

\item{w_factor}{Water content (g m-2). Typical value 100 (0-500).}

\item{l_factor}{Lignin and cellulose content. Combined lignin and cellulose
content (mg m-2). Typical value 40 (10-80).}

\item{p_factor}{Nitrogen content (g m-2). Typical value 1  (0.3-2.0).}
}
\value{
Length 4 list containing modeled reflectance, transmittance,
wavelengths, and R (?).
}
\description{
LIBERTY model
}
\author{
Alexey Shiklomanov
}
