% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prosail.R
\name{pro4sail_4}
\alias{pro4sail_4}
\alias{pro4sail_5}
\alias{pro4sail_d}
\title{PRO4SAIL: PROSPECT leaf model coupled to 4SAIL canopy RTM and Hapke soil
model}
\usage{
pro4sail_4(N, Cab, Cw, Cm, LAI, soil_moisture, ...)

pro4sail_5(N, Cab, Car, Cbrown, Cw, Cm, LAI, soil_moisture, ...)

pro4sail_d(N, Cab, Car, Canth, Cbrown, Cw, Cm, LAI, soil_moisture, ...)
}
\arguments{
\item{N}{Effective number of leaf layers}

\item{Cab}{Chlorophyll content (ug cm -2)}

\item{Cw}{Leaf water content (g cm-2), or leaf water thickness (cm)}

\item{Cm}{Dry matter content (g cm-2)}

\item{LAI}{Leaf area index}

\item{soil_moisture}{Soil moisture index (0-1; 0 = completely dry,
1 = completely wet)}

\item{...}{Additional arguments to \link{foursail}}

\item{Car}{Carotenoid content (ug cm -2)}

\item{Cbrown}{"Brown matter" content (unitless)}

\item{Canth}{Anthocyanin content (ug cm-2)}
}
\value{
List containing four reflectance streams: bi-hemispherical (diffuse
in, diffuse out; BHR), directional-hemispherical (direct in, diffuse out;
DHR), hemispherical-directional (diffuse in, direct out; HDR), and
bi-directional (direct in, direct out; BHR).
}
\description{
PRO4SAIL: PROSPECT leaf model coupled to 4SAIL canopy RTM and Hapke soil
model
}
\author{
Alexey Shiklomanov
}
