% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospect.R
\name{prospect4}
\alias{prospect4}
\alias{prospect5}
\alias{prospectd}
\title{PROSPECT leaf spectra model}
\usage{
prospect4(N, Cab, Cw, Cm)

prospect5(N, Cab, Car, Cw, Cm, Cbrown = 0)

prospectd(N, Cab, Car, Canth, Cbrown, Cw, Cm)
}
\arguments{
\item{N}{Effective number of leaf layers}

\item{Cab}{Chlorophyll content (ug cm -2)}

\item{Cw}{Leaf water content (g cm-2), or leaf water thickness (cm)}

\item{Cm}{Dry matter content (g cm-2)}

\item{Car}{Carotenoid content (ug cm -2)}

\item{Cbrown}{"Brown matter" content (unitless)}

\item{Canth}{Anthocyanin content (ug cm-2)}
}
\value{
Matrix of reflectance and transmittance values
}
\description{
PROSPECT leaf spectra model
}
\author{
Alexey Shiklomanov
}
