% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataspec.R
\docType{data}
\name{refractive_p45}
\alias{refractive_p45}
\alias{refractive_pd}
\title{Refractive index for PROSPECT model}
\format{An object of class \code{numeric} of length 2101.}
\usage{
refractive_p45

refractive_pd
}
\description{
Refractive index for PROSPECT model
}
\keyword{datasets}
