% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral-response.R
\docType{data}
\name{sensor_list}
\alias{sensor_list}
\title{Sensor spectral response functions}
\format{
An object of class \code{character} of length 12.
}
\usage{
sensor_list
}
\description{
Sensor spectral response functions
}
\keyword{datasets}
