% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edr-allometry.R
\name{size2bl}
\alias{size2bl}
\title{Calculate leaf biomass given DBH}
\usage{
size2bl(dbh, b1, b2)
}
\arguments{
\item{dbh}{Cohort diameter at breast height}

\item{b1}{Allometry equation base}

\item{b2}{Allometry equation exponent}
}
\value{
Leaf biomass
}
\description{
Allometry equation is of the form \code{Y = b1 * DBH ^ b2}.
}
\details{
Original function in ED src is in \code{utils/allometry.f90}. See also
the \code{area_indices} subroutine in the same file.

LAI is calculated as \code{bleaf * nplant * SLA}.
}
\author{
Alexey Shiklomanov
}
