% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{solvearray}
\alias{solvearray}
\title{Solve the matrix equation AX = B for X for an array of matrices A}
\usage{
solvearray(a, b)
}
\arguments{
\item{a}{3D array. For dimensions x,y,z, iteration is over dimension z ("slices")}

\item{b}{Matrix with dimensions y,x}
}
\value{
Matrix of solutions (x,z)
}
\description{
Solve the matrix equation AX = B for X for an array of matrices A
}
