% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral-response.R
\name{spectral_response}
\alias{spectral_response}
\title{Convolution of spectra to sensor RSR}
\usage{
spectral_response(spec, sensor)
}
\arguments{
\item{spec}{Full (1 nm) spectrum (vector)}

\item{sensor}{Sensor name (string). See \code{sensor_list}.}
}
\description{
Convolution of spectra to sensor RSR
}
