% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edr.R
\name{sw_two_stream}
\alias{sw_two_stream}
\title{R implementation of ED2 two-stream radiation model}
\usage{
sw_two_stream(
  czen,
  iota_g,
  pft,
  lai,
  wai,
  cai,
  orient_factor,
  clumping_factor,
  leaf_reflect,
  leaf_trans,
  wood_reflect,
  wood_trans,
  down_sky,
  down0_sky,
  wavelengths = seq(400, 2500)
)
}
\arguments{
\item{czen}{Cosine of angle of incidence (\code{cosaio})}

\item{iota_g}{Ground (soil + snow) albedo (nwl)}

\item{pft}{PFT identities of each cohort, as integer (ncoh)}

\item{lai}{Leaf area index of each cohort (ncoh)}

\item{wai}{Wood area index of each cohort (ncoh)}

\item{cai}{Crown area of each cohort (ncoh)}

\item{orient_factor}{Orient factor (npft)}

\item{clumping_factor}{Clumping factor (npft)}

\item{leaf_reflect}{Leaf reflectance spectra (nwl * npft)}

\item{leaf_trans}{Leaf transmittance spectra (nwl * npft)}

\item{wood_reflect}{Wood reflectance spectra (nwl * npft)}

\item{wood_trans}{Wood transmittance spectra (nwl * npft)}

\item{down_sky}{Normalized diffuse solar spectrum (nwl)}

\item{down0_sky}{Normalized direct solar spectrum (nwl)}

\item{wavelengths}{Numeric vector of wavelengths to use, in nm
(nwl). Default is 400:2500.}
}
\value{

}
\description{
R implementation of ED2 two-stream radiation model
}
\author{
Alexey Shiklomanov
}
