% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpm.R
\name{tav_abs}
\alias{tav_abs}
\title{\code{tav} function}
\usage{
tav_abs(theta, refractive)
}
\arguments{
\item{theta}{Angle of incident radiation (in degrees)}

\item{refractive}{Refractive index (vector)}
}
\description{
\code{tav} function
}
\references{
Stern F. (1964), Transmission of isotropic radiation
across an interface between two dielectrics, Appl. Opt.,
3(1):111-113. Allen W.A. (1973), Transmission of isotropic light
across a dielectric surface in two and three dimensions, J. Opt.
Soc. Am., 63(6):664-666.
}
