% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sail-volscatt.R
\name{volscatt}
\alias{volscatt}
\title{Volume scattering function}
\usage{
volscatt(solar_zenith, instrument_zenith, azimuth, leaf_angle)
}
\arguments{
\item{solar_zenith}{Solar zenith angle (degrees)}

\item{instrument_zenith}{Instrument zenith angle (degrees)}

\item{azimuth}{Sun-instrument azimuth angle (degrees)}

\item{leaf_angle}{Leaf angle (degrees)}
}
\description{
Volume scattering function
}
