% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sail-volscatt.R
\name{volscatt_scalar}
\alias{volscatt_scalar}
\title{Original volscatt implementation, based on SAIL Fortran code. Exists for
testing of faster vectorized version.}
\usage{
volscatt_scalar(solar_zenith, instrument_zenith, azimuth, leaf_angle)
}
\arguments{
\item{solar_zenith}{Solar zenith angle (degrees)}

\item{instrument_zenith}{Instrument zenith angle (degrees)}

\item{azimuth}{Sun-instrument azimuth angle (degrees)}

\item{leaf_angle}{Leaf angle (degrees)}
}
\description{
Original volscatt implementation, based on SAIL Fortran code. Exists for
testing of faster vectorized version.
}
\author{
Alexey Shiklomanov
}
