% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sail-volscatt.R
\name{volscatt_scalar_v}
\alias{volscatt_scalar_v}
\title{"Vectorized" version of scalar code}
\usage{
volscatt_scalar_v(solar_zenith, instrument_zenith, azimuth, leaf_angle)
}
\arguments{
\item{solar_zenith}{Solar zenith angle (degrees)}

\item{instrument_zenith}{Instrument zenith angle (degrees)}

\item{azimuth}{Sun-instrument azimuth angle (degrees)}

\item{leaf_angle}{Vector of leaf angles (degrees)}
}
\description{
"Vectorized" version of scalar code
}
