% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edr-allometry.R
\name{wai_allometry}
\alias{wai_allometry}
\title{Wood allometry}
\usage{
wai_allometry(dbh, nplant, b1, b2)
}
\arguments{
\item{dbh}{Cohort diameter at breast height}

\item{nplant}{Stem density (stems m-2)}

\item{b1}{Allometry equation base}

\item{b2}{Allometry equation exponent}
}
\value{
Wood area index
}
\description{
Calculate wood area index (WAI) given DBH
}
\details{
Two possibilities here, depending on \code{iallom}.
The default value is: `WAI = nplant * b1 * DBH ^ b2

An alternative formulation is that it is always 11\% of LAI: \code{WAI = 0.11 * LAI}. But we don't use that here because it's trivial.
}
\author{
Alexey Shiklomanov
}
