% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liberty.R
\name{liberty}
\alias{liberty}
\title{LIBERTY model}
\usage{
liberty(
  D,
  xu,
  thick,
  baseline,
  element,
  c_factor,
  w_factor,
  l_factor,
  p_factor
)
}
\arguments{
\item{D}{Cell diameter. Average leaf cell diameter (m-6). Typical value 40 (20-100).}

\item{xu}{Intercellular air space. Determinant for the amount of radiative
flux passing between cells. Typical value 0.045 (0.01-0.1)}

\item{thick}{Leaf thickness. Arbitrary value to determine single leaf
reflectance and transmittance from infinite reflectance criteria. Typical
value 1.6 (1-10).}

\item{baseline}{Baseline absorption. Wavelength-independent absorption to
compensate for changes in absolute reflectance. Typical value 0.0006 for
fresh leaves or or 0.0004 for dry leaves.}

\item{element}{Albino absorption. Absorption in the visible region due to
lignin. Typical value 2 (0-4).}

\item{c_factor}{Chlorophyll content (mg m-2). Typical value 200 (0-600).}

\item{w_factor}{Water content (g m-2). Typical value 100 (0-500).}

\item{l_factor}{Lignin and cellulose content. Combined lignin and cellulose
content (mg m-2). Typical value 40 (10-80).}

\item{p_factor}{Nitrogen content (g m-2). Typical value 1  (0.3-2.0).}
}
\value{
Length 4 list containing modeled reflectance, transmittance,
wavelengths, and R (?).
}
\description{
Leaf radiative transfer model designed for conifer needles.
}
\details{
Note that these coefficients are based on the original values from Dawson et
al. (1998). Improved specific absorption coefficients are available from
later work by Di Vittorio (2009).
}
\references{
Dawson, T. P., Curran, P. J., & Plummer, S. E. (1998). LIBERTY—Modeling the
Effects of Leaf Biochemical Concentration on Reflectance Spectra. Remote
Sensing of Environment, 65(1), 50–60.
https://doi.org/10.1016/S0034-4257(98)00007-8

Di Vittorio, A. V. (2009). Enhancing a leaf radiative transfer
model to estimate concentrations and in vivo specific absorption
coefficients of total carotenoids and chlorophylls a and b from
single-needle reflectance and transmittance. Remote Sensing of Environment,
113(9), 1948–1966. https://doi.org/10.1016/j.rse.2009.05.002
}
\author{
Alexey Shiklomanov
}
