% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sail-volscatt.R
\name{volscatt_scalar_v}
\alias{volscatt_scalar_v}
\title{"Vectorized" version of scalar code}
\usage{
volscatt_scalar_v(solar_zenith, instrument_zenith, azimuth, leaf_angle)
}
\description{
"Vectorized" version of scalar code
}
